backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain
langchain-openai
langchain-community
pypdf
faiss-cpu # Use faiss-gpu if you have a compatible GPU
python-multipart